pyproject.toml
[tool.poetry]
name = "photobot"
version = "0.1.0"
description = ""
authors = ["kiortir <mpgg2000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.0.2"
pydantic = "^2.1.1"
aiogram = "^2.25.1"
pillow = "^10.0.0"
httpx = "^0.24.1"
aiofiles = "^23.1.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
